+++ lib/puppet_x/puppetlabs/azure/provider_arm.rb
@@ -60,10 +60,11 @@ module PuppetX
             params = build_params(args)
 
             if ! args[:managed_disks]
-              create_storage_account({
-                storage_account: args[:storage_account],
+              # ensure the storage account is set up
+              get_or_create_storage_account({
+                name: args[:storage_account],
                 resource_group: args[:resource_group],
-                storage_account_type: args[:storage_account_type],
+                sku_name: args[:storage_account_type],
                 location: args[:location],
                 tags: args[:tags],
               })
@@ -182,6 +183,21 @@ module PuppetX
           end
         end
 
+        def get_all_vnets # rubocop:disable Metrics/AbcSize
+          begin
+            Puppet.debug "Getting all virtual networks from subscription"
+            ProviderArm.network_client.virtual_networks.list_all.collect do |vnet|
+              ProviderArm.network_client.virtual_networks.get(resource_group_from(vnet), vnet.name)
+            end
+          rescue MsRestAzure::AzureOperationError => err
+            raise Puppet::Error, JSON.parse(err.message)['message']
+          rescue MsRest::DeserializationError => err
+            raise Puppet::Error, err.response_body
+          rescue MsRest::RestError => err
+            raise Puppet::Error, err.to_s
+          end
+        end
+
         def get_all_vms # rubocop:disable Metrics/AbcSize
           begin
             vms = []
@@ -291,14 +307,49 @@ module PuppetX
           end
         end
 
-        def create_storage_account(args)
-          params = build_storage_account_create_parameters(args)
-          ProviderArm.storage_client.storage_accounts.create(args[:resource_group], args[:storage_account], params)
+        def get_or_create_storage_account(args) # rubocop:disable Metrics/AbcSize
+          # ensure the storage account exists
+          begin
+            get_storage_account({
+              name: args[:name],
+              resource_group: args[:resource_group],
+            })
+          rescue MsRestAzure::AzureOperationError
+            create_storage_account({
+              name: args[:name],
+              resource_group: args[:resource_group],
+              sku_name: args[:sku_name],
+              location: args[:location],
+            })
+          rescue MsRestAzure::AzureOperationError => err
+            raise Puppet::Error, JSON.parse(err.message)['message']
+          rescue MsRest::DeserializationError => err
+            raise Puppet::Error, err.response_body
+          rescue MsRest::RestError => err
+            raise Puppet::Error, err.to_s
+          end
         end
 
-        def delete_storage_account(sa)
+        def get_storage_account(args)
+          ProviderArm.storage_client.storage_accounts.get_properties(args[:resource_group], args[:name])
+        end
+
+        def create_storage_account(args)
+          params = build_storage_account_create_parameters(args)
           begin
-            ProviderArm.storage_client.storage_accounts.delete(resource_group, sa.name)
+            ProviderArm.storage_client.storage_accounts.create(args[:resource_group], args[:name], params)
+          rescue MsRestAzure::AzureOperationError => err
+            raise Puppet::Error, JSON.parse(err.message)['message']
+          rescue MsRest::DeserializationError => err
+            raise Puppet::Error, err.response_body
+          rescue MsRest::RestError => err
+            raise Puppet::Error, err.to_s
+          end
+        end
+
+        def delete_storage_account(args)
+          begin
+            ProviderArm.storage_client.storage_accounts.delete(args[:resource_group], args[:name])
           rescue MsRestAzure::AzureOperationError => err
             raise Puppet::Error, JSON.parse(err.message)['message']
           rescue MsRest::DeserializationError => err
@@ -327,20 +378,51 @@ module PuppetX
           end
         end
 
-        def retrieve_virtual_network(args)
+        def retrieve_virtual_network(args) # rubocop:disable Metrics/AbcSize
           begin
             ProviderArm.network_client.virtual_networks.get(
               *expand_name([args[:resource_group]], args[:virtual_network_name]),
             )
           rescue MsRestAzure::AzureOperationError
-            create_virtual_network(args)
+            # Auto-create a specified but missing virtual network
+            create_virtual_network({
+              resource_group: args[:resource_group],
+              name: args[:virtual_network_name],
+              location: args[:location],
+              address_prefixes: [args[:virtual_network_address_space]],
+              dns_servers: args[:dns_servers].split,
+            }) unless args[:location].nil? || args[:virtual_network_address_space].nil?
+
+            # Auto-create the subnet
+            create_subnet(args)
           end
         end
 
-        def create_virtual_network(args)
-          Puppet.debug("Creating vnet '#{args[:virtual_network_name]}'")
-          params = build_virtual_network_params(args)
-          ProviderArm.network_client.virtual_networks.create_or_update(args[:resource_group], args[:virtual_network_name], params)
+        def create_virtual_network(args) # rubocop:disable Metrics/AbcSize
+          Puppet.debug("Creating vnet '#{args[:resource_group]}/#{args[:name]}'")
+          begin
+            params = build_virtual_network_params(args)
+            ProviderArm.network_client.virtual_networks.create_or_update(args[:resource_group], args[:name], params)
+          rescue MsRestAzure::AzureOperationError => err
+            raise Puppet::Error, JSON.parse(err.message)['message']
+          rescue MsRest::DeserializationError => err
+            raise Puppet::Error, err.response_body
+          rescue MsRest::RestError => err
+            raise Puppet::Error, err.to_s
+          end
+        end
+
+        def delete_virtual_network(rg, name)
+          Puppet.debug("Deleting vnet '#{rg}/#{name}'")
+          begin
+            ProviderArm.network_client.virtual_networks.delete(rg, name)
+          rescue MsRestAzure::AzureOperationError => err
+            raise Puppet::Error, JSON.parse(err.message)['message']
+          rescue MsRest::DeserializationError => err
+            raise Puppet::Error, err.response_body
+          rescue MsRest::RestError => err
+            raise Puppet::Error, err.to_s
+          end
         end
 
         def create_public_ip_address(args)
@@ -356,22 +438,22 @@ module PuppetX
           ProviderArm.network_client.public_ipaddresses.get(resource_group, public_ip_address_name)
         end
 
-        def retrieve_subnet(virtual_network, args)
+        def retrieve_subnet(args)
           begin
             ProviderArm.network_client.subnets.get(
               *expand_name([args[:resource_group], args[:virtual_network_name]], args[:subnet_name]),
             )
           rescue MsRestAzure::AzureOperationError
-            create_subnet(virtual_network, args)
+            create_subnet(args)
           end
         end
 
-        def create_subnet(virtual_network, args)
-          Puppet.debug("Creating subnet '#{args[:subnet_name]}' on vnet '#{virtual_network.name}'")
+        def create_subnet(args)
+          Puppet.debug("Creating subnet '#{args[:subnet_name]}' on vnet '#{args[:virtual_network_name]}'")
           params = build_subnet_params(args)
           ProviderArm.network_client.subnets.create_or_update(
             args[:resource_group],
-            virtual_network.name,
+            args[:virtual_network_name],
             args[:subnet_name],
             params
           )
@@ -457,14 +539,18 @@ module PuppetX
         end
 
         def build_storage_account_create_parameters(args)
-          build(::Azure::ARM::Storage::Models::StorageAccountCreateParameters, {
+          buildparams = {
             location: args[:location],
             tags: args[:tags],
-            kind: Object.const_get("::Azure::ARM::Storage::Models::Kind::#{args[:storage_account_kind] || :Storage}"),
+            kind: Object.const_get("::Azure::ARM::Storage::Models::Kind::#{args[:account_kind] || :Storage}"),
             sku: build(::Azure::ARM::Storage::Models::Sku, {
-              name: args[:storage_account_type],
+              name: args[:sku_name],
             })
-          })
+          }
+          if args[:account_kind] == :BlobStorage
+            buildparams[:access_tier] = Object.const_get("::Azure::ARM::Storage::Models::AccessTier::#{args[:access_tier] || :Hot}")
+          end
+          build(::Azure::ARM::Storage::Models::StorageAccountCreateParameters, buildparams)
         end
 
         def build_virtual_machine_extensions(args) # rubocop:disable Metrics/AbcSize
@@ -489,30 +575,39 @@ module PuppetX
           })
         end
 
-        def build_storage_profile(args)
-          os_disk =
-            if args[:managed_disks]
-              build(::Azure::ARM::Compute::Models::OSDisk, {
-                create_option: 'FromImage',
-                managed_disk: build(::Azure::ARM::Compute::Models::ManagedDiskParameters, {
-                  storage_account_type: args[:storage_account_type],
-                }),
-              })
-            else
-              build(::Azure::ARM::Compute::Models::OSDisk, {
-                caching: args[:os_disk_caching],
-                create_option: args[:os_disk_create_option],
-                name: args[:os_disk_name],
-                vhd: build(::Azure::ARM::Compute::Models::VirtualHardDisk, {
-                  uri: build_os_vhd_uri(args),
+        def build_storage_profile(args) # rubocop:disable Metrics/AbcSize
+          begin
+            os_disk =
+              if args[:managed_disks]
+                build(::Azure::ARM::Compute::Models::OSDisk, {
+                  create_option: 'FromImage',
+                  managed_disk: build(::Azure::ARM::Compute::Models::ManagedDiskParameters, {
+                    storage_account_type: args[:storage_account_type],
+                  }),
                 })
-              })
-            end
-          build(::Azure::ARM::Compute::Models::StorageProfile, {
-            image_reference: build_image_reference(args),
-            os_disk: os_disk,
-            data_disks: build_data_disks(args),
-          })
+              else
+                build(::Azure::ARM::Compute::Models::OSDisk, {
+                  caching: args[:os_disk_caching],
+                  create_option: args[:os_disk_create_option],
+                  name: args[:os_disk_name],
+                  vhd: build(::Azure::ARM::Compute::Models::VirtualHardDisk, {
+                    uri: build_os_vhd_uri(args),
+                  }),
+                })
+              end
+
+            build(::Azure::ARM::Compute::Models::StorageProfile, {
+              image_reference: build_image_reference(args),
+              os_disk: os_disk,
+              data_disks: build_data_disks(args),
+            })
+          rescue MsRestAzure::AzureOperationError => err
+            raise Puppet::Error, JSON.parse(err.message)['message']
+          rescue MsRest::DeserializationError => err
+            raise Puppet::Error, err.response_body
+          rescue MsRest::RestError => err
+            raise Puppet::Error, err.to_s
+          end
         end
 
         def build_data_disks(args)
@@ -545,20 +640,16 @@ module PuppetX
         end
 
         def build_virtual_network_params(args)
-          build(::Azure::ARM::Network::Models::VirtualNetwork, {
+          vnet_params = {
             location: args[:location],
             address_space: build(::Azure::ARM::Network::Models::AddressSpace, {
-              address_prefixes: args[:virtual_network_address_space],
+              address_prefixes: args[:address_prefixes],
             }),
             dhcp_options: build(::Azure::ARM::Network::Models::DhcpOptions, {
-              dns_servers: args[:dns_servers].split,
+              dns_servers: args[:dns_servers],
             }),
-            #XXX This should handle arrays
-            subnets: [build(::Azure::ARM::Network::Models::Subnet, {
-              name: args[:subnet_name],
-              address_prefix: args[:subnet_address_prefix],
-            })]
-          })
+          }
+          build(::Azure::ARM::Network::Models::VirtualNetwork, vnet_params)
         end
 
         def build_subnet_params(args)
@@ -605,7 +696,7 @@ module PuppetX
             network_interfaces: [
               create_network_interface(
                 args,
-                retrieve_subnet(retrieve_virtual_network(args), args),
+                retrieve_subnet(args),
               )
             ]
           })