+++ spec/acceptance/vnet_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper_acceptance'
 
-describe 'azure_storage_account when creating a storage account' do
+describe 'azure_vnet when creating a virtual network' do
   include_context 'with a known name and storage account name'
   include_context 'destroy left-over created ARM resources after use'
 
@@ -10,40 +10,40 @@ describe 'azure_storage_account when creating a storage account' do
     @config = {
       name: @name,
       ensure: 'present',
+      location: CHEAPEST_ARM_LOCATION,
+      resource_group: SPEC_RESOURCE_GROUP,
       optional: {
-        location: CHEAPEST_ARM_LOCATION,
-        resource_group: SPEC_RESOURCE_GROUP,
-        sku_name: 'Standard_GRS',
-        account_kind: 'Storage',
+      },
+      nonstring: {
+        address_prefixes: ['172.16.0.0/12'],
+        dns_servers: ['8.8.8.8'],
       },
     }
-    @template = 'azure_storage_account.pp.tmpl'
+    @template = 'azure_vnet.pp.tmpl'
     @manifest = PuppetManifest.new(@template, @config)
     @result = @manifest.execute
-    @machine = @client.get_storage_account(@name)
+    @vnet = @client.get_vnet(@name)
   end
 
   it_behaves_like 'an idempotent resource'
 
   it 'should have the correct name' do
-    expect(@machine.name).to eq(@name)
+    expect(@vnet.name).to eq(@name)
   end
 
   context 'when puppet resource is run' do
-    include_context 'a puppet ARM resource run', 'azure_storage_account'
+    include_context 'a puppet ARM resource run', 'azure_vnet'
     puppet_resource_should_show('ensure', 'present')
     puppet_resource_should_show('location', CHEAPEST_ARM_LOCATION)
-    puppet_resource_should_show('sku_name', 'Standard_GRS')
-    puppet_resource_should_show('account_kind', 'Storage')
     puppet_resource_should_show('resource_group', SPEC_RESOURCE_GROUP.downcase)
   end
 
-  context 'when we try and destroy the SA' do
+  context 'when we try and destroy the vnet' do
     before(:all) do
       new_config = @config.update({:ensure => 'absent'})
       manifest = PuppetManifest.new(@template, new_config)
       @result = manifest.execute
-      @machine = @client.get_storage_account(@name)
+      @vnet = @client.get_vnet(@name)
     end
 
     it 'should run without errors' do
@@ -51,7 +51,7 @@ describe 'azure_storage_account when creating a storage account' do
     end
 
     it 'should be destroyed' do
-      expect(@machine).to be_nil
+      expect(@vnet).to be_nil
     end
   end
 end