+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+
+## Supported Version 1.2.2
+This release adds the azure_vnet resource type
+
+## Supported Version 1.2.1
+This release adds support for blob, cold, and https-only storage accounts
+
 ## Supported Version 1.2.0
 ### Added
  - Support for Azure's managed disks feature removes the requirement to associate a storage account with each Azure VM, removing one of the fundamental limitations of the platform.